+++ src/context-store--indexable/modifiers/get-update-one-context-data.test.ts
@@ -153,7 +153,7 @@ describe(getTestName(__dirname), () => {
         },
         state: "success",
       };
-      return expect(
+      await expect(
         getUpdateOneContextData(
           statefulIndexStore,
           setContextData,