+++ src/context-store--stateful-indexable/interfaces.ts
@@ -1,18 +1,23 @@
-import { IndexableContextStore } from "../context-store--indexable/interfaces";
 import { ContextStore } from "../context-store--basic";
 
-export type IndexableStatefulContextStore<TDataItem> = IndexableContextStore<
-  ContextStore<TDataItem>
->;
+// Helper type for when data is an indexable type (map or array)
+export type IndexableContextStoreData<IndexableContextStoreItem> =
+  | Record<number | string | symbol, IndexableContextStoreItem>
+  | Array<IndexableContextStoreItem>;
+
+export interface IndexableStatefulContextStore<TDataItem>
+  extends ContextStore<IndexableContextStoreData<ContextStore<TDataItem>>> {}
 
 export type IndexableStatefulContextStoreValue<
-  TContextStore extends IndexableStatefulContextStore<any>
+  TContextStore extends IndexableStatefulContextStore<unknown>
 > = TContextStore["data"][any];
 
 export type IndexableStatefulContextStoreValueData<
-  TContextStore extends IndexableStatefulContextStore<any>
+  TContextStore extends IndexableStatefulContextStore<unknown>
 > = TContextStore["data"][any]["data"];
 
 export type IndexableStatefulContextStoreKey<
-  TContextStore extends IndexableStatefulContextStore<any>
-> = keyof TContextStore["data"];
+  TContextStore extends IndexableStatefulContextStore<unknown>
+> = TContextStore["data"] extends Array<any>
+  ? number
+  : keyof TContextStore["data"];